Revert "Upgrade from Chromium 89 to Chromium 90."

+++ browser/component_updater/brave_component_updater_configurator.cc
@@ -57,7 +57,7 @@ BraveConfigurator::BraveConfigurator(
 
 BraveConfigurator::~BraveConfigurator() {}
 
-double BraveConfigurator::InitialDelay() const {
+int BraveConfigurator::InitialDelay() const {
   return configurator_impl_.InitialDelay();
 }
